requirements.txt
pyotp~=2.9.0
qrcode~=8.2
django-qrcode~=0.3
pillow~=11.3.0